Extract Product::Release class
