+++ lib/end_of_life/ruby_version.rb
@@ -2,71 +2,43 @@ require "net/http"
 require "rubygems"
 
 module EndOfLife
-  class RubyVersion
-    include Comparable
+  module RubyVersion
+    extend self
+    include Dry::Monads[:try]
 
-    ZERO = Gem::Version.new("0")
-
-    class << self
-      include Dry::Monads[:try]
-
-      def from_file(file_name:, content:, parser: Parser)
-        parser.parse_file(file_name: file_name, content: content)
-      end
-
-      def eol_versions_at(date)
-        all_versions.filter { |version| version.eol_date <= date }
-      end
-
-      def latest_eol(at: Date.today)
-        eol_versions_at(at).max
-      end
-
-      private
-
-      EOL_API_URL = "https://endoflife.date/api/ruby.json"
-      DB_PATH = File.join(__dir__, "../end_of_life.json")
-
-      def all_versions
-        @all_versions ||= (fetch_end_of_life_api || load_file_fallback)
-          .then { |json| JSON.parse(json, symbolize_names: true) }
-          .map { |version| new(version[:latest], eol_date: Date.parse(version[:eol])) }
-      end
-
-      def fetch_end_of_life_api
-        Try { Net::HTTP.get URI(EOL_API_URL) }.value_or(nil)
-      end
-
-      def load_file_fallback
-        File.read(DB_PATH)
-      end
+    def from_file(file_name:, content:, parser: Parser)
+      parser.parse_file(file_name: file_name, content: content)
     end
 
-    attr_reader :version, :eol_date
-
-    def initialize(version_string, eol_date: nil)
-      @version = Gem::Version.new(version_string)
-      @eol_date = eol_date
-
-      freeze
+    def eol_versions_at(date)
+      all_versions.filter { |version| version.eol_date <= date }
     end
 
-    def eol?(at: Date.today)
-      if eol_date
-        eol_date <= at
-      else
-        self <= RubyVersion.latest_eol(at: at)
-      end
+    def latest_eol(at: Date.today)
+      eol_versions_at(at).max
     end
 
-    def <=>(other)
-      @version <=> other.version
+    def new(version_string, eol_date: nil)
+      Product::Release.new(product: :ruby, version: version_string, eol_date: eol_date)
     end
 
-    def zero? = @version == ZERO
+    private
 
-    def to_s
-      @version.to_s
+    EOL_API_URL = "https://endoflife.date/api/ruby.json"
+    DB_PATH = File.join(__dir__, "../end_of_life.json")
+
+    def all_versions
+      @all_versions ||= fetch_end_of_life_api.value_or(load_file_fallback)
+        .then { |json| JSON.parse(json, symbolize_names: true) }
+        .map { |version| new(version[:latest], eol_date: Date.parse(version[:eol])) }
+    end
+
+    def fetch_end_of_life_api
+      Try { Net::HTTP.get URI(EOL_API_URL) }
+    end
+
+    def load_file_fallback
+      File.read(DB_PATH)
     end
   end
 end
